Move HSLuv and HPLuv tests to hsluv directory
There are a large number of tests in the HSLuv and HPLuv test
suites and there's no need to run them everytime so they have
been moved out of the primary test suite into their own test
directory.
