+++ test/hsluv/hpluv.js
@@ -1,6 +1,6 @@
-import { to, sRGB, HPLuv } from "../src/index-fn.js";
-import { check } from "./util.mjs";
-import { readTestData, normalizeCoords } from "./hsluv/util.mjs";
+import { to, sRGB, HPLuv } from "../../src/index-fn.js";
+import { check } from "../util.mjs";
+import { readTestData, normalizeCoords } from "./util.mjs";
 
 let json = readTestData();
 let srgbToHpluv = [];
